requirements.txt
chromadb==0.5.17
discord.py==2.4.0
langchain==0.3.6
langchain_community==0.3.4
langchain_core==0.3.15
langchain_openai==0.2.5
python-dotenv==1.0.1
